requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
fastapi
uvicorn
python-dotenv
pypdf
openai
groq
